+++ app-backend/api/src/main/scala/campaign/Routes.scala
@@ -62,6 +62,13 @@ trait CampaignRoutes
               }
             }
           } ~
+          pathPrefix("retrieve-child-labels") {
+            pathEndOrSingleSlash {
+              post {
+                retrieveChildCampaignLabels(campaignId)
+              }
+            }
+          } ~
           pathPrefix("permissions") {
             pathEndOrSingleSlash {
               get {
@@ -366,5 +373,27 @@ trait CampaignRoutes
         }
       }
     }
+  }
+
+  def retrieveChildCampaignLabels(campaignId: UUID): Route = authenticate {
+    user =>
+      authorizeScope(
+        ScopedAction(Domain.Campaigns, Action.Clone, None),
+        user
+      ) {
+        authorizeAuthResultAsync {
+          CampaignDao
+            .authorized(user, ObjectType.Campaign, campaignId, ActionType.Edit)
+            .transact(xa)
+            .unsafeToFuture
+        } {
+          onSuccess {
+            CampaignDao
+              .retrieveChildCampaignAnnotations(campaignId)
+              .transact(xa)
+              .unsafeToFuture
+          } { complete(StatusCodes.NoContent) }
+        }
+      }
   }
 }