+++ app-backend/db/src/main/scala/AnnotationLabelClassDao.scala
@@ -3,6 +3,7 @@ package com.rasterfoundry.database
 import com.rasterfoundry.database.Implicits._
 import com.rasterfoundry.datamodel._
 
+import cats.implicits._
 import doobie._
 import doobie.implicits._
 import doobie.postgres.implicits._
@@ -28,15 +29,20 @@ object AnnotationLabelClassDao extends Dao[AnnotationLabelClass] {
 
   def insertAnnotationLabelClass(
       classCreate: AnnotationLabelClass.Create,
-      annotationLabelGroup: AnnotationLabelClassGroup
-  ): ConnectionIO[AnnotationLabelClass] = {
-    (fr"INSERT INTO" ++ tableF ++ fr"(" ++ insertFieldsF ++ fr")" ++
-      fr"""VALUES (
+      annotationLabelGroup: AnnotationLabelClassGroup,
+      parent: Option[AnnotationLabelClass]
+  ): ConnectionIO[AnnotationLabelClass] =
+    for {
+      newLabelClass <- (fr"INSERT INTO" ++ tableF ++ fr"(" ++ insertFieldsF ++ fr")" ++
+        fr"""VALUES (
         uuid_generate_v4(), ${classCreate.name}, ${annotationLabelGroup.id},
         ${classCreate.colorHexCode}, ${classCreate.default}, ${classCreate.determinant},
         ${classCreate.index}, ${classCreate.geometryType}, ${classCreate.description}
       )""").update.withUniqueGeneratedKeys[AnnotationLabelClass](fieldNames: _*)
-  }
+      _ <- parent traverse { parentClass =>
+        fr"INSERT INTO label_class_history VALUES (${parentClass.id}, ${newLabelClass.id})".update.run
+      }
+    } yield newLabelClass
 
   def listAnnotationLabelClassByGroupId(
       groupId: UUID